`envelope.rb`: Cleaning this file
